[cartographer_ros] Do not drop data when falling behind. (#68)
Dropping data significantly harms quality of SLAM. This change means that
all data will be kept. If the system falls behind only for a short period
of time this is preferable. Publishing of the pose and submaps does not
work in this case and needs improvement.

Original commit:
googlecartographer/cartographer_ros@9103b846ffb6213fa2fe5b2507494ae0113c8f9b