[cartographer] Remove HybridGrid slicing into a ProbabilityGrid. (#385)
This code is currently unused. The feature of slicing a HybridGrid into a Probability grid could be useful to for example collect a map in 3D, then slice the 3D data to create a 2D map. There are simpler ways to implement this, especially in the context of the assets_writer_main.

PAIR=wohe

Original commit:
googlecartographer/cartographer@611f244d7ac269ff4a39811b7baa8f7b3bea2a29